refactor: split stragety to folder, rename _originFieldX/Y to originFieldX/Y, remove defaultState params
